simple-django/requirement.txt
Django==1.11
MySQL-python==1.2.5
virtualenv==15.1.0

# Django REST Framework
djangorestframework==3.7.7

# Celery with Redis
amqp==2.2.2
redis==2.10.6
celery==4.1.0
billiard==3.5.0.3
kombu==4.1.0
pytz==2017.2
vine==1.1.4

# Coverage
coverage==4.4.1

# Code Highlighting
Pygments==2.2.0

# MarkDown
Markdown==2.6.10

# pytest and pytest-django
attrs==17.3.0
funcsigs==1.0.2
pluggy==0.6.0
py==1.5.2
pytest==3.3.1
pytest-django==3.1.2
six==1.11.0

# Coreapi and dependencies
certifi==2017.11.5
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
idna==2.6
itypes==1.1.0
Jinja2==2.10
MarkupSafe==1.0
requests==2.18.4
uritemplate==3.0.0
urllib3==1.22
# Coreapi_cli
click==6.7
coreapi-cli==1.0.6